refactor: centralize execution time logic into shared utility functions for consistency
